update with new owloop functions name
