+++ CMakeLists.txt
@@ -9,7 +9,7 @@ project(sit_pkg)
 # Catkin
 ##############################################################################
 
-find_package(catkin REQUIRED multi_ontology_reference owloop_pkg rosjava_build_tools) #ros_java_msgs
+#find_package(catkin REQUIRED multi_ontology_reference owloop_pkg rosjava_build_tools) #ros_java_msgs
 
 # Set the gradle targets you want catkin's make to run by default, e.g.
 #   catkin_rosjava_setup(installApp)
