+++ sit/main/java/it/emarolab/sit/owloopDescriptor/SceneIndividualDescriptor.java
@@ -1,12 +1,12 @@
-package it.emarolab.scene_identification_tagging.owloopDescriptor;
+package it.emarolab.sit.owloopDescriptor;
 
 import it.emarolab.amor.owlInterface.OWLReferences;
-import it.emarolab.owloop.aMORDescriptor.MORAxioms;
-import it.emarolab.owloop.aMORDescriptor.MORIndividual;
-import it.emarolab.owloop.aMORDescriptor.utility.MORIndividualBase;
-import it.emarolab.owloop.aMORDescriptor.utility.dataProperty.MORHierarchicalDataProperty;
-import it.emarolab.owloop.aMORDescriptor.utility.objectProperty.MORHierarchicalObjectProperty;
-import it.emarolab.scene_identification_tagging.SITBase;
+import it.emarolab.owloop.descriptor.construction.descriptorEntitySet.DescriptorEntitySet;
+import it.emarolab.owloop.descriptor.construction.descriptorExpression.IndividualExpression;
+import it.emarolab.owloop.descriptor.construction.descriptorGround.IndividualGround;
+import it.emarolab.owloop.descriptor.utility.dataPropertyDescriptor.HierarchicalDataPropertyDesc;
+import it.emarolab.owloop.descriptor.utility.objectPropertyDescriptor.HierarchicalObjectPropertyDesc;
+import it.emarolab.sit.SITBase;
 import org.semanticweb.owlapi.model.OWLClass;
 import org.semanticweb.owlapi.model.OWLNamedIndividual;
 
@@ -22,7 +22,7 @@ import java.util.List;
  *     the scene cardinality associate to a concrete scene representation.
  *
  * <div style="text-align:center;"><small>
- * <b>File</b>:        it.emarolab.scene_identification_tagging.owloopDescriptor.SceneIndividualDescriptor <br>
+ * <b>File</b>:        it.emarolab.sit.owloopDescriptor.SceneIndividualDescriptor <br>
  * <b>Licence</b>:     GNU GENERAL PUBLIC LICENSE. Version 3, 29 June 2007 <br>
  * <b>Author</b>:      Buoncompagni Luca (luca.buoncompagni@edu.unige.it) <br>
  * <b>affiliation</b>: EMAROLab, DIBRIS, University of Genoa. <br>
@@ -30,17 +30,17 @@ import java.util.List;
  * </small></div>
  */
 public class SceneIndividualDescriptor
-        extends MORIndividualBase
-        implements MORIndividual.Type<SceneClassDescriptor>,
-        MORIndividual.ObjectLink<MORHierarchicalObjectProperty>,
-        MORIndividual.DataLink<MORHierarchicalDataProperty>,
+        extends IndividualGround
+        implements IndividualExpression.Type<SceneClassDescriptor>,
+        IndividualExpression.ObjectLink<HierarchicalObjectPropertyDesc>,
+        IndividualExpression.DataLink<HierarchicalDataPropertyDesc>,
         SITBase{
 
     private int cardinality = 0;
 
-    private MORAxioms.Concepts individualTypes = new MORAxioms.Concepts();
-    private MORAxioms.ObjectSemantics objectLinks = new MORAxioms.ObjectSemantics();
-    private MORAxioms.DataSemantics dataLinks = new MORAxioms.DataSemantics();
+    private DescriptorEntitySet.Concepts individualTypes = new DescriptorEntitySet.Concepts();
+    private DescriptorEntitySet.ObjectLinksSet objectLinks = new DescriptorEntitySet.ObjectLinksSet();
+    private DescriptorEntitySet.DataLinksSet dataLinks = new DescriptorEntitySet.DataLinksSet();
 
 
     /**
@@ -82,10 +82,10 @@ public class SceneIndividualDescriptor
      * @return the changes done during reading.
      */
     @Override
-    public List<MappingIntent> readSemantic() {
-        List<MappingIntent> r = MORIndividual.Type.super.readSemantic();
-        r.addAll( MORIndividual.ObjectLink.super.readSemantic());
-        r.addAll( MORIndividual.DataLink.super.readSemantic());
+    public List<MappingIntent> readExpressionAxioms() {
+        List<MappingIntent> r = IndividualExpression.Type.super.readExpressionAxioms();
+        r.addAll( IndividualExpression.ObjectLink.super.readExpressionAxioms());
+        r.addAll( IndividualExpression.DataLink.super.readExpressionAxioms());
         updateCardinality();
         return r;
     }
@@ -96,78 +96,25 @@ public class SceneIndividualDescriptor
      * @return the changes done during writing.
      */
     @Override
-    public List<MappingIntent> writeSemantic() {
-        List<MappingIntent> r = MORIndividual.Type.super.writeSemantic();
-        r.addAll( MORIndividual.ObjectLink.super.writeSemantic());
-        r.addAll( MORIndividual.DataLink.super.writeSemantic());
+    public List<MappingIntent> writeExpressionAxioms() {
+        List<MappingIntent> r = IndividualExpression.Type.super.writeExpressionAxioms();
+        r.addAll( IndividualExpression.ObjectLink.super.writeExpressionAxioms());
+        r.addAll( IndividualExpression.DataLink.super.writeExpressionAxioms());
         updateCardinality();
         return r;
     }
 
     private void updateCardinality(){
         cardinality = 0; // hp: all object properties are spatial relations
-        for ( MORAxioms.ObjectSemantic s : getObjectSemantics())
+        for ( DescriptorEntitySet.ObjectLinks s : getIndividualObjectProperties())
             cardinality += s.getValues().size();
-    }
-
-    /**
-     * This is a standard OWLOOP implementation.
-     * @return a new {@code Descriptor} for a scene class in which {@code this} individual belongs.
-     */
-    @Override // you can change the returning type to any implementations of MORConcept
-    public SceneClassDescriptor getNewTypeIndividual(OWLClass instance, OWLReferences ontology) {
-        return new SceneClassDescriptor( instance, ontology);
-    }
-
-    /**
-     * This is a standard OWLOOP implementation.
-     * @return the name of all the scene classes in which {@code this} individual belongs.
-     */
-    @Override
-    public MORAxioms.Concepts getTypeIndividual() {
-        return individualTypes;
-    }
-
-    /**
-     * This is a standard OWLOOP implementation.
-     * @return an object property of {@code this} individual.
-     */
-    @Override
-    public MORHierarchicalObjectProperty getNewObjectIndividual(MORAxioms.ObjectSemantic instance, OWLReferences ontology) {
-        return new MORHierarchicalObjectProperty( instance.getSemantic(), ontology);
-    }
-
-    /**
-     * This is a standard OWLOOP implementation.
-     * @return the names and values of all the synchronised object properties of {@code this} individual.
-     */
-    @Override
-    public MORAxioms.ObjectSemantics getObjectSemantics() {
-        return objectLinks;
-    }
-
-    /**
-     * This is a standard OWLOOP implementation.
-     * @return a data property of {@code this} individual.
-     */
-    @Override
-    public MORHierarchicalDataProperty getNewDataIndividual(MORAxioms.DataSemantic instance, OWLReferences ontology) {
-        return new MORHierarchicalDataProperty( instance.getSemantic(), ontology);
-    }
-
-    /**
-     * This is a standard OWLOOP implementation.
-     * @return the names and values of all the synchronised data properties of {@code this} individual.
-     */
-    @Override
-    public MORAxioms.DataSemantics getDataSemantics() {
-        return dataLinks;
+        cardinality=cardinality/2;
     }
 
     /**
      * The sum of all the spatial object properties applied to {@code this}
      * concrete scene individual.
-     * @return the individual cardinality based on {@link #readSemantic()} and {@link #writeSemantic()}.
+     * @return the individual cardinality based on {@link #readExpressionAxioms()} ()} and {@link #writeExpressionAxioms()} ()}.
      */
     public int getCardinality() {
         return cardinality;
@@ -181,5 +128,61 @@ public class SceneIndividualDescriptor
                 "," + NL + "\t⊨ " + objectLinks +
                 "," + NL + "\t⊢ " + dataLinks +
                 NL + "}";
+    }
+
+    /**
+     * This is a standard OWLOOP implementation.
+     * @return a new {@code Descriptor} for a scene class in which {@code this} individual belongs.
+     */
+    @Override
+    public SceneClassDescriptor getNewIndividualType(OWLClass instance, OWLReferences ontology) {
+        return new SceneClassDescriptor( instance, ontology);
+    }
+
+
+    /**
+     * This is a standard OWLOOP implementation.
+     * @return an object property of {@code this} individual.
+     */
+    @Override
+    public HierarchicalObjectPropertyDesc getNewIndividualObjectProperty(DescriptorEntitySet.ObjectLinks instance, OWLReferences ontology) {
+        return new HierarchicalObjectPropertyDesc( instance.getExpression(), ontology);
+    }
+
+    /**
+     * This is a standard OWLOOP implementation.
+     * @return a data property of {@code this} individual.
+     */
+    @Override
+    public HierarchicalDataPropertyDesc getNewIndividualDataProperty(DescriptorEntitySet.DataLinks instance, OWLReferences ontology) {
+        return new HierarchicalDataPropertyDesc( instance.getExpression(), ontology);
+    }
+
+    /**
+     * This is a standard OWLOOP implementation.
+     * @return the names and values of all the synchronised data properties of {@code this} individual.
+     */
+    @Override
+    public DescriptorEntitySet.DataLinksSet getIndividualDataProperties() {
+        return dataLinks;
+    }
+
+    /**
+     * This is a standard OWLOOP implementation.
+     * @return the names and values of all the synchronised object properties of {@code this} individual.
+     */
+    @Override
+    public DescriptorEntitySet.ObjectLinksSet getIndividualObjectProperties() {
+        return objectLinks;
+    }
+
+
+    /**
+     * This is a standard OWLOOP implementation.
+     * @return the name of all the scene classes in which {@code this} individual belongs.
+     */
+    @Override
+    public DescriptorEntitySet.Concepts getIndividualTypes() {
+        return individualTypes;
     }
 }
