+++ sit/main/java/it/emarolab/sit/realObject/Cone.java
@@ -1,53 +1,53 @@
-package it.emarolab.scene_identification_tagging.realObject;
+package it.emarolab.sit.realObject;
 
 import it.emarolab.amor.owlInterface.OWLReferences;
-import it.emarolab.scene_identification_tagging.SITBase;
-import it.emarolab.scene_identification_tagging.owloopDescriptor.SpatialIndividualDescriptor;
+import it.emarolab.sit.SITBase;
+import it.emarolab.sit.owloopDescriptor.SpatialIndividualDescriptor;
 import org.semanticweb.owlapi.model.OWLNamedIndividual;
 
 import java.util.List;
 
 /**
- * The ontological cylinder descriptor.
+ * The ontological cone descriptor.
  * <p>
  *     This class describes extends the basic {@link Orientable}
- *     by considering also a generic point in the axis, a radius and a height, in accord
+ *     by considering also a apex, a radius and a height, in accord
  *     with the ontological structure.
  *
  * <div style="text-align:center;"><small>
- * <b>File</b>:        it.emarolab.scene_identification_tagging.realObject.Cylinder <br>
+ * <b>File</b>:        it.emarolab.sit.realObject.Cone <br>
  * <b>Licence</b>:     GNU GENERAL PUBLIC LICENSE. Version 3, 29 June 2007 <br>
  * <b>Author</b>:      Buoncompagni Luca (luca.buoncompagni@edu.unige.it) <br>
  * <b>affiliation</b>: EMAROLab, DIBRIS, University of Genoa. <br>
  * <b>date</b>:        05/06/17 <br>
  * </small></div>
  */
-public class Cylinder
+public class Cone
         extends Orientable
         implements SITBase{
 
-    private Float pointX = null;
-    private Float pointY = null;
-    private Float pointZ = null;
+    private Float apexX = null;
+    private Float apexY = null;
+    private Float apexZ = null;
     private Float height = null;
     private Float radius = null;
 
     /**
      * Initialise and ontological object by have a name based on {@link #currentCOUNT}
-     * and {@link INDIVIDUAL#PREFIX_CYLINDER}.
+     * and {@link INDIVIDUAL#PREFIX_CONE}.
      * @param onto the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Cylinder(OWLReferences onto) {
-        super( INDIVIDUAL.PREFIX_CYLINDER + currentCOUNT, onto);
+    public Cone(OWLReferences onto) {
+        super( INDIVIDUAL.PREFIX_CONE + currentCOUNT, onto);
         initialiseProperty();
     }
     /**
      * Initialise and ontological object by have a name based on {@link #currentCOUNT}
-     * and {@link INDIVIDUAL#PREFIX_CYLINDER}.
+     * and {@link INDIVIDUAL#PREFIX_CONE}.
      * @param ontoName the name of the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Cylinder(String ontoName) {
-        super(INDIVIDUAL.PREFIX_CYLINDER + currentCOUNT, ontoName);
+    public Cone(String ontoName) {
+        super(INDIVIDUAL.PREFIX_CONE + currentCOUNT, ontoName);
         initialiseProperty();
     }
     /**
@@ -55,7 +55,7 @@ public class Cylinder
      * @param instance the object individual.
      * @param onto the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Cylinder(OWLNamedIndividual instance, OWLReferences onto) {
+    public Cone(OWLNamedIndividual instance, OWLReferences onto) {
         super(instance, onto);
         initialiseProperty();
     }
@@ -64,7 +64,7 @@ public class Cylinder
      * @param instanceName the name of the object individual.
      * @param onto the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Cylinder(String instanceName, OWLReferences onto) {
+    public Cone(String instanceName, OWLReferences onto) {
         super(instanceName, onto);
         initialiseProperty();
     }
@@ -73,7 +73,7 @@ public class Cylinder
      * @param instance the object individual.
      * @param ontoName the name of the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Cylinder(OWLNamedIndividual instance, String ontoName) {
+    public Cone(OWLNamedIndividual instance, String ontoName) {
         super(instance, ontoName);
         initialiseProperty();
     }
@@ -82,157 +82,151 @@ public class Cylinder
      * @param instanceName the name of the object individual.
      * @param ontoName the name of the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Cylinder(String instanceName, String ontoName) {
+    public Cone(String instanceName, String ontoName) {
         super(instanceName, ontoName);
         initialiseProperty();
     }
 
     // common constructor
     private void initialiseProperty(){
-        addData( getPropertyPointX(), true);
-        addData( getPropertyPointY(), true);
-        addData( getPropertyPointZ(), true);
+        addData( getPropertyApexX(), true);
+        addData( getPropertyApexY(), true);
+        addData( getPropertyApexZ(), true);
         addData( getPropertyHeight(), true);
         addData( getPropertyRadius(), true);
-        addTypeIndividual( CLASS.CYLINDER);
+        addTypeIndividual( CLASS.CONE);
     }
 
     /**
      * Return the data property used to map the {@code X}
-     * coordinate of a generic point in the cylinder principal axis.
-     * @return {@link DATA_PROPERTY#CYLINDER_POINT_X}
+     * coordinate of the apex of the code.
+     * @return {@link DATA_PROPERTY#CONE_APEX_X}
      */
-    public String getPropertyPointX(){
-        return DATA_PROPERTY.CYLINDER_POINT_X;
+    public String getPropertyApexX(){
+        return DATA_PROPERTY.CONE_APEX_X;
     }
 
     /**
      * Return the data property used to map the {@code Y}
-     * coordinate of a generic point in the cylinder principal axis.
-     * @return {@link DATA_PROPERTY#CYLINDER_POINT_Y}
+     * coordinate of the apex of the code.
+     * @return {@link DATA_PROPERTY#CONE_APEX_Y}
      */
-    public String getPropertyPointY(){
-        return DATA_PROPERTY.CYLINDER_POINT_Y;
+    public String getPropertyApexY(){
+        return DATA_PROPERTY.CONE_APEX_Y;
     }
 
     /**
      * Return the data property used to map the {@code Z}
-     * coordinate of a generic point in the cylinder principal axis.
-     * @return {@link DATA_PROPERTY#CYLINDER_POINT_Z}
+     * coordinate of the apex of the code.
+     * @return {@link DATA_PROPERTY#CONE_APEX_Z}
      */
-    public String getPropertyPointZ(){
-        return DATA_PROPERTY.CYLINDER_POINT_Z;
+    public String getPropertyApexZ(){
+        return DATA_PROPERTY.CONE_APEX_Z;
     }
 
     /**
-     * Return the data property used to map the height of the cylinder.
-     * @return {@link DATA_PROPERTY#CYLINDER_HEIGHT}
+     * Return the data property used to map the height of the code.
+     * @return {@link DATA_PROPERTY#CONE_HEIGHT}
      */
     public String getPropertyHeight(){
-        return DATA_PROPERTY.CYLINDER_HEIGHT;
+        return DATA_PROPERTY.CONE_HEIGHT;
     }
 
     /**
-     * Return the data property used to map the height of the cylinder.
-     * @return {@link DATA_PROPERTY#CYLINDER_RADIUS}
+     * Return the data property used to map the radius of the cone.
+     * @return {@link DATA_PROPERTY#CONE_RADIUS}
      */
     public String getPropertyRadius(){
-        return DATA_PROPERTY.CYLINDER_RADIUS;
+        return DATA_PROPERTY.CONE_RADIUS;
     }
 
     /**
-     * A shortcut for calling {@link #setPointX(Float)},
-     * {@link #setPointY(Float)} and {@link #setPointZ(Float)}.
-     * @param x the {@code X} coordinate of a point in the axis.
-     * @param y the {@code Y} coordinate of a point in the axis.
-     * @param z the {@code Z} coordinate of a point in the axis.
+     * A shortcut for calling {@link #setApexX(Float)},
+     * {@link #setApexY(Float)} and {@link #setApexZ(Float)}.
+     * @param x the {@code X} coordinate of the apex.
+     * @param y the {@code Y} coordinate of the apex.
+     * @param z the {@code Z} coordinate of the apex.
      */
     public void setApex(Float x, Float y, Float z){
-        setPointX( x);
-        setPointY( y);
-        setPointZ( z);
+        setApexX( x);
+        setApexY( y);
+        setApexZ( z);
     }
 
     /**
-     * Returns the {@code X} coordinate of a
-     * point belonging to the principal axis.
+     * Returns the {@code X} coordinate of the apex.
      * Measurements should be in meters.
-     * @return the {@code X} coordinate of a point in the axis.
+     * @return the {@code X} component of the apex.
      */
-    public Float getPointX() {
-        return pointX;
+    public Float getApexX() {
+        return apexX;
     }
 
     /**
-     * Set the {@code X} coordinate of
-     * a pointin the axis.
+     * Set the {@code X} coordinate of the apex.
      * Measurements should be in meters.
      * This method automatically add the data to the
      * {@link SpatialIndividualDescriptor}.
-     * @param x the {@code X} coordinate of a point in the axis.
+     * @param x the {@code X} coordinate of the apex.
      */
-    public void setPointX(Float x) {
-        this.pointX = x;
-        addData( getPropertyPointX(), x, true);
+    public void setApexX(Float x) {
+        this.apexX = x;
+        addData( getPropertyApexX(), x, true);
     }
 
     /**
-     * Returns the {@code Y} coordinate of a
-     * point belonging to the principal axis.
+     * Returns the {@code Y} coordinate of the apex.
      * Measurements should be in meters.
-     * @return the {@code Y} coordinate of a point in the axis.
+     * @return the {@code Y} component of the apex.
      */
-    public Float getPointY() {
-        return pointY;
+    public Float getApexY() {
+        return apexY;
     }
 
     /**
-     * Set the {@code Y} coordinate of
-     * a pointin the axis.
+     * Set the {@code Y} coordinate of the apex.
      * Measurements should be in meters.
      * This method automatically add the data to the
      * {@link SpatialIndividualDescriptor}.
-     * @param y the {@code Y} coordinate of a point in the axis.
+     * @param y the {@code Y} coordinate of the apex.
      */
-    public void setPointY(Float y) {
-        this.pointY = y;
-        addData( getPropertyPointY(), y, true);
+    public void setApexY(Float y) {
+        this.apexY = y;
+        addData( getPropertyApexY(), y, true);
     }
 
     /**
-     * Returns the {@code Z} coordinate of a
-     * point belonging to the principal axis.
+     * Returns the {@code Z} coordinate of the apex.
      * Measurements should be in meters.
-     * @return the {@code Z} coordinate of a point in the axis.
+     * @return the {@code Z} component of the apex.
      */
-    public Float getPointZ() {
-        return pointZ;
+    public Float getApexZ() {
+        return apexZ;
     }
 
     /**
-     * Set the {@code Z} coordinate of
-     * a pointin the axis.
+     * Set the {@code Z} coordinate of the apex.
      * Measurements should be in meters.
      * This method automatically add the data to the
      * {@link SpatialIndividualDescriptor}.
-     * @param z the {@code Z} coordinate of a point in the axis.
+     * @param z the {@code Z} coordinate of the apex.
      */
-    public void setPointZ(Float z) {
-        this.pointZ = z;
-        addData( getPropertyPointZ(), z, true);
+    public void setApexZ(Float z) {
+        this.apexZ = z;
+        addData( getPropertyApexZ(), z, true);
     }
 
     /**
-     * Returns the height value of the cylinder.
+     * Returns the height value of the cone.
      * Measurements should be in meters.
-     * @return the height.
+     * @return the height of the cone.
      */
     public Float getHight() {
         return height;
     }
 
     /**
-     * Set the height value of the cylinder.
+     * Set the height value of the cone.
      * Measurements should be in meters.
      * This method automatically add the data to the
      * {@link SpatialIndividualDescriptor}.
@@ -244,16 +238,16 @@ public class Cylinder
     }
 
     /**
-     * Returns the radius value of the cylinder.
+     * Returns the radius value of the cone.
      * Measurements should be in meters.
-     * @return the radius.
+     * @return the radius of the cone.
      */
     public Float getRadius() {
         return radius;
     }
 
     /**
-     * Set the radius value of the cylinder.
+     * Set the radius value of the cone.
      * Measurements should be in meters.
      * This method automatically add the data to the
      * {@link SpatialIndividualDescriptor}.
@@ -266,16 +260,16 @@ public class Cylinder
 
     /**
      * Enhance the standard OWLOOP read semantic by
-     * explicitly set the value of a point in the axis, the height and the radius
+     * explicitly set the value of the apex, the height and the radius
      * as well as other measurements from super classes.
      * @return the changes done during reading.
      */
     @Override
-    public List<MappingIntent> readSemantic() {
-        List<MappingIntent> r = super.readSemantic();
-        pointX = getLiteral( getPropertyPointX()).parseFloat();
-        pointY = getLiteral( getPropertyPointY()).parseFloat();
-        pointZ = getLiteral( getPropertyPointZ()).parseFloat();
+    public List<MappingIntent> readExpressionAxioms() {
+        List<MappingIntent> r = super.readExpressionAxioms();
+        apexX = getLiteral( getPropertyApexX()).parseFloat();
+        apexY = getLiteral( getPropertyApexY()).parseFloat();
+        apexZ = getLiteral( getPropertyApexZ()).parseFloat();
         height = getLiteral( getPropertyHeight()).parseFloat();
         radius = getLiteral( getPropertyRadius()).parseFloat();
         return r;
