+++ sit/main/java/it/emarolab/sit/realObject/Plane.java
@@ -1,49 +1,49 @@
-package it.emarolab.scene_identification_tagging.realObject;
+package it.emarolab.sit.realObject;
 
 import it.emarolab.amor.owlInterface.OWLReferences;
-import it.emarolab.scene_identification_tagging.SITBase;
-import it.emarolab.scene_identification_tagging.owloopDescriptor.SpatialIndividualDescriptor;
+import it.emarolab.sit.SITBase;
+import it.emarolab.sit.owloopDescriptor.SpatialIndividualDescriptor;
 import org.semanticweb.owlapi.model.OWLNamedIndividual;
 
 import java.util.List;
 
 /**
- * The ontological sphere descriptor.
+ * The ontological plane descriptor.
  * <p>
- *     This class describes extends the basic {@link GeometricPrimitive}
- *     by considering also a radius for a sphere, in accord
+ *     This class describes extends the basic {@link Orientable}
+ *     by considering also an hessian for a plane, in accord
  *     with the ontological structure.
  *
  * <div style="text-align:center;"><small>
- * <b>File</b>:        it.emarolab.scene_identification_tagging.realObject.Sphere <br>
+ * <b>File</b>:        it.emarolab.sit.realObject.Plane <br>
  * <b>Licence</b>:     GNU GENERAL PUBLIC LICENSE. Version 3, 29 June 2007 <br>
  * <b>Author</b>:      Buoncompagni Luca (luca.buoncompagni@edu.unige.it) <br>
  * <b>affiliation</b>: EMAROLab, DIBRIS, University of Genoa. <br>
  * <b>date</b>:        05/06/17 <br>
  * </small></div>
  */
-public class Sphere
-        extends GeometricPrimitive
+public class Plane
+        extends Orientable
         implements SITBase{
 
-    private Float radius = null;
+    private Float hessian = null;
 
     /**
      * Initialise and ontological object by have a name based on {@link #currentCOUNT}
-     * and {@link INDIVIDUAL#PREFIX_SPHERE}.
+     * and {@link INDIVIDUAL#PREFIX_PLANE}.
      * @param onto the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Sphere(OWLReferences onto) {
-        super( INDIVIDUAL.PREFIX_SPHERE + currentCOUNT, onto);
+    public Plane(OWLReferences onto) {
+        super( INDIVIDUAL.PREFIX_PLANE + currentCOUNT, onto);
         initialiseProperty();
     }
     /**
      * Initialise and ontological object by have a name based on {@link #currentCOUNT}
-     * and {@link INDIVIDUAL#PREFIX_SPHERE}.
+     * and {@link INDIVIDUAL#PREFIX_PLANE}.
      * @param ontoName the name of the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Sphere(String ontoName) {
-        super(INDIVIDUAL.PREFIX_SPHERE + currentCOUNT, ontoName);
+    public Plane(String ontoName) {
+        super(INDIVIDUAL.PREFIX_PLANE + currentCOUNT, ontoName);
         initialiseProperty();
     }
     /**
@@ -51,7 +51,7 @@ public class Sphere
      * @param instance the object individual.
      * @param onto the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Sphere(OWLNamedIndividual instance, OWLReferences onto) {
+    public Plane(OWLNamedIndividual instance, OWLReferences onto) {
         super(instance, onto);
         initialiseProperty();
     }
@@ -60,7 +60,7 @@ public class Sphere
      * @param instanceName the name of the object individual.
      * @param onto the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Sphere(String instanceName, OWLReferences onto) {
+    public Plane(String instanceName, OWLReferences onto) {
         super(instanceName, onto);
         initialiseProperty();
     }
@@ -69,7 +69,7 @@ public class Sphere
      * @param instance the object individual.
      * @param ontoName the name of the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Sphere(OWLNamedIndividual instance, String ontoName) {
+    public Plane(OWLNamedIndividual instance, String ontoName) {
         super(instance, ontoName);
         initialiseProperty();
     }
@@ -78,56 +78,56 @@ public class Sphere
      * @param instanceName the name of the object individual.
      * @param ontoName the name of the {@link OWLReferences} ontology that will contain the object individual.
      */
-    public Sphere(String instanceName, String ontoName) {
+    public Plane(String instanceName, String ontoName) {
         super(instanceName, ontoName);
         initialiseProperty();
     }
 
     // common constructor
     private void initialiseProperty(){
-        addData( getPropertyRadius(), true);
-        addTypeIndividual( CLASS.SPHERE);
+        addData( getPropertyHessian(), true);
+        addTypeIndividual( CLASS.PLANE);
     }
 
     /**
-     * Return the name of data property used to map the radius.
-     * @return {@link DATA_PROPERTY#RADIUS_SPHERE}
+     * Return the name of data property used to map the hessian.
+     * @return {@link DATA_PROPERTY#HESSIAN}
      */
-    public String getPropertyRadius(){
-        return DATA_PROPERTY.RADIUS_SPHERE;
+    public String getPropertyHessian(){
+        return DATA_PROPERTY.HESSIAN;
     }
 
     /**
-     * Returns the radius of this sphere
+     * Returns the hessian of this plane
      * Measurements should be in meters.
-     * @return the sphere radius.
+     * @return the plane hessian.
      */
-    public Float getRadius() {
-        return radius;
+    public Float getHessian() {
+        return hessian;
     }
 
     /**
-     * Set the radius of the sphere.
+     * Set the hessian of the plane.
      * Measurements should be in meters.
      * This method automatically add the data to the
      * {@link SpatialIndividualDescriptor}.
-     * @param r the radius of the sphere to set.
+     * @param h the hessian of the plane to set.
      */
-    public void setRadius(Float r) {
-        this.radius = r;
-        addData( getPropertyRadius(), r, true);
+    public void setHessian(Float h) {
+        this.hessian = h;
+        addData( getPropertyHessian(), h, true);
     }
 
     /**
      * Enhance the standard OWLOOP read semantic by
-     * explicitly set the value of the radius,
+     * explicitly set the value of the hessian,
      * as well as other measurements from super classes.
      * @return the changes done during reading.
      */
     @Override
-    public List<MappingIntent> readSemantic() {
-        List<MappingIntent> r = super.readSemantic();
-        radius = getLiteral( getPropertyRadius()).parseFloat();
+    public List<MappingIntent> readExpressionAxioms() {
+        List<MappingIntent> r = super.readExpressionAxioms();
+        hessian = getLiteral( getPropertyHessian()).parseFloat();
         return r;
     }
 }
